Add utils function
- Chunk function was repeated in several components, moved to shared js
